requirements.txt
streamlit>=1.30
openpyxl>=3.1
pandas>=2.0
anthropic>=0.40
python-docx>=1.0
openai>=1.30
google-genai>=1.0
pdfplumber>=0.10
